requirement.txt
fastapi==0.58.1
starlette==0.13.4
python-dotenv==0.13.0
uvicorn==0.11.5
telethon==1.15.0
pydantic==1.5.1
